+++ db/migrate/20200721061437_create_participation_communities.rb
@@ -1,8 +1,8 @@
 class CreateParticipationCommunities < ActiveRecord::Migration[6.0]
   def change
     create_table :participation_communities do |t|
-      t.integer :user_id
-      t.integer :community_id
+      t.bigint :user_id
+      t.bigint :community_id
 
       t.timestamps
     end
